changes the extra questions for
